Delete bogus claim about unit name starting with a digit
Checked with Per and John. None of us can remember where that worry
started and I can't find the relevant emails anymore. There should be
no common sense reason why the name of a unit of measure would start
with a digit, and I found no examples of such units.